+++ indispensable-asn1/src/commonMain/kotlin/at/asitplus/signum/indispensable/asn1/Encoding.kt
@@ -48,17 +48,3 @@ internal fun Sink.appendUnsafe(bytes: ByteArray, startIndex: Int = 0, endIndex: 
     }
     return endIndex - startIndex
 }
-
-/**
- * Drops bytes at the start, or adds zero bytes at the start, until the [size] is reached
- */
-fun ByteArray.ensureSize(size: Int): ByteArray = (this.size - size).let { toDrop ->
-    when {
-        toDrop > 0 -> this.copyOfRange(toDrop, this.size)
-        toDrop < 0 -> ByteArray(-toDrop) + this
-        else -> this
-    }
-}
-
-@Suppress("NOTHING_TO_INLINE")
-inline fun ByteArray.ensureSize(size: UInt) = ensureSize(size.toInt())