+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Better safe than sorry!
 * Directly support UUID-based OID creation
 * Implement hash-to-curve and hash-to-scalar as per RFC9380
 * Rename `decodeFromDerHexString` to `parseFromDerHexString`
+* Move `ensureSize` from package `asn1` to `misc`
 * `JwsSignes.plainsignatureInput` is now a raw ByteArray
   * `JwsSigned.prepareSignatureInput` now returns a raw ByteArray
 * Use kotlinx-io as primary source for parsing